extract base class for local repo operations
